requirements.txt
cryptography
pycryptodomex
mmh3
bitarray
pynacl